refactor (BallStatus): ⚙️ enums 로 이동
