+++ src/test/java/study/baseballRemaster/model/BallTest.java
@@ -5,8 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import study.baseballRemaster.model.Ball;
-import study.baseballRemaster.model.BallStatus;
+import study.baseballRemaster.enums.BallStatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
