+++ src/test/java/study/baseballRemaster/model/BallsTest.java
@@ -3,11 +3,7 @@ package study.baseballRemaster.model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import study.baseballRemaster.model.Ball;
-import study.baseballRemaster.model.BallStatus;
-import study.baseballRemaster.model.Balls;
-import study.baseballRemaster.model.Judge;
-import study.baseballRemaster.model.Player;
+import study.baseballRemaster.enums.BallStatus;
 
 import java.util.Arrays;
